+++ ui/library/src/utils/VisibleEntriesDetector.ts
